Add UserApi.addFavorite and test deserialization
This commit introduces the `addFavorite` function to the `UserApi`. This function allows adding a recipe to a user's favorites.

A corresponding test, `addFavorite should deserialize correctly`, has been added to `UserApiTest.kt` to ensure the response is deserialized as `Unit` when the API call is successful (indicated by an empty JSON response).