Add delete recipe endpoint
Adds the ability to delete a recipe by its slug or ID.

This is achieved by introducing a `deleteRecipe` function to the `RecipesApi` interface, which makes a `DELETE` request to the `api/recipes/{slug}` endpoint.

Additionally, a corresponding unit test has been added to verify that the new endpoint returns a successful response.