Add duplicate recipe endpoint (#83)
This commit introduces the functionality to duplicate an existing
recipe.

A `duplicateRecipe` suspend function has been added to the `RecipesApi`
interface. This function sends a `POST` request to the
`api/recipes/{slug}/duplicate` endpoint to create a copy of the
specified recipe.

A corresponding test, `duplicateRecipe should deserialize correctly`,
has been added to verify that the response is deserialized correctly.