Add endpoint to update a recipe's `lastMade` date (#84)
This commit introduces the ability to update the `lastMade` date for a
specific recipe.

A new `updateLastMade` suspend function has been added to the
`RecipesApi` interface. This function sends a `PATCH` request to the
`recipes/{slug}/last-made` endpoint to set the new timestamp.

To support this, the following changes were made:
- A new data class, `RecipeLastMadeJson`, was created to model the
request body, containing a single `timestamp` field.
- A corresponding unit test, `updateLastMade should return unit`, was
added to verify that the new endpoint call executes successfully and
handles an empty success response.