Add endpoint to scrape a recipe image from a URL
This commit introduces functionality to scrape an image from a given URL and assign it as the image for a specific recipe.

A new `scrapeImageUrl` suspend function has been added to the `RecipesApi` interface. This function sends a `POST` request to the `recipes/{slug}/image` endpoint.

To support this, the following changes were made:
- A new data class, `ScrapeImageUrlRequestJson`, was created to model the request body, which includes the `url` of the image to scrape and an `includeTags` flag.
- A corresponding unit test, `scrapeImageUrl should construct url and deserialize response correctly`, was added to verify that the request is constructed correctly and that an empty success response is handled properly.