Add endpoint to update a recipe's image
This commit introduces the functionality to update the image for a specific recipe.

A new `updateImage` suspend function has been added to the `RecipesApi` interface. This function sends a `PUT` request to the `recipes/{slug}/image` endpoint to upload a new image for the recipe identified by its slug.

To support this, the following changes were made:
- The `UpdateRecipeImageRequest` data class was created to model the request body, which includes the base64-encoded image and its file extension.
- The `UpdateRecipeImageResponse` data class was created to handle the response, which contains the filename of the new image.