Add endpoint to upload a recipe asset (#89)
This commit introduces the functionality to upload an asset for a
specific recipe.

A new `uploadAsset` suspend function has been added to the `RecipesApi`
interface. This function sends a `POST` request to the
`recipes/{slug}/assets` endpoint to upload the asset file.

To support this, the following changes were made:
- The `UploadAssetRequestJson` data class was created to model the
request body, which includes the asset's name, icon, extension, and the
base64 encoded file content.
- The `UploadAssetResponseJson` data class was created to model the
successful response from the server.
- A new unit test, `uploadAsset should construct url and deserialize
response correctly`, was added to verify that the request URL is
correctly constructed and that the response is deserialized properly.